同步初始提交

+++ app/src/main/java/com/botongsoft/rfid/bean/classity/CheckError.java
@@ -27,7 +27,7 @@ public class CheckError implements java.io.Serializable {
     @Column
     private int status;
     @Column
-    private int anchor;
+    private long anchor;
 
     public int getStatus() {
         return status;
@@ -37,11 +37,11 @@ public class CheckError implements java.io.Serializable {
         this.status = status;
     }
 
-    public int getAnchor() {
+    public long getAnchor() {
         return anchor;
     }
 
-    public void setAnchor(int anchor) {
+    public void setAnchor(long anchor) {
         this.anchor = anchor;
     }
