+++ app/src/main/java/com/botongsoft/rfid/bean/classity/Kf.java
@@ -65,7 +65,7 @@ public class Kf implements java.io.Serializable {
     @Column
     private int status;
     @Column
-    private int anchor;
+    private long anchor;
 
     public int getStatus() {
         return status;
@@ -75,11 +75,11 @@ public class Kf implements java.io.Serializable {
         this.status = status;
     }
 
-    public int getAnchor() {
+    public long getAnchor() {
         return anchor;
     }
 
-    public void setAnchor(int anchor) {
+    public void setAnchor(long anchor) {
         this.anchor = anchor;
     }
 }
