+++ app/src/main/java/com/botongsoft/rfid/bean/classity/Mjj.java
@@ -42,7 +42,7 @@ public class Mjj implements java.io.Serializable {
     @Column
     private int status;
     @Column
-    private int anchor;
+    private long anchor;
 
     public int getStatus() {
         return status;
@@ -52,11 +52,11 @@ public class Mjj implements java.io.Serializable {
         this.status = status;
     }
 
-    public int getAnchor() {
+    public long getAnchor() {
         return anchor;
     }
 
-    public void setAnchor(int anchor) {
+    public void setAnchor(long anchor) {
         this.anchor = anchor;
     }
